Files that conform the repository
The idea is to have a graphic interface that allows to understand how the bubble sort algorithm works. The code is made by three files following the MVC approach. Where the main class is the graphic display, the Bubble class is the Model and the Controller links the two of them.